refactor(nat): Move NatPort checked type to its own submodule
Leave room for the IP/port allocator in allocator.rs, move NatPort
definitions and implementations to their own, dedicated file.
