+++ nat/src/stateful/mod.rs
@@ -6,6 +6,7 @@
 #![allow(unused_imports)]
 
 mod allocator;
+mod port;
 pub mod sessions;
 
 use crate::stateful::sessions::{
@@ -149,7 +150,7 @@ impl StatefulNat {
     fn set_source_port(
         transport: &mut Transport,
         next_header: NextHeader,
-        target_port: Option<allocator::NatPort>,
+        target_port: Option<port::NatPort>,
     ) {
         let Some(port) = target_port else {
             return;
@@ -168,7 +169,7 @@ impl StatefulNat {
     fn set_destination_port(
         transport: &mut Transport,
         next_header: NextHeader,
-        target_port: Option<allocator::NatPort>,
+        target_port: Option<port::NatPort>,
     ) {
         let Some(port) = target_port else {
             return;
@@ -295,7 +296,7 @@ impl<Buf: PacketBufferMut> NetworkFunction<Buf> for StatefulNat {
 
 #[cfg(test)]
 mod tests {
-    use super::allocator::NatPort;
+    use super::port::NatPort;
     use super::*;
     use net::packet::test_utils::build_test_ipv4_packet;
     use net::tcp::Tcp;
